+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## Unreleased
+### Changed
+* Updated google dependencies and made required changes
+
 ## [1.1.5] - 2017-08-31
 ### Added
 * Support for longitude named `long`